[Bug #157930395]Modified the file structure and tests for API endpoints
