--- app/.gitignore
@@ -1,4 +1,0 @@
-.vscode
-venv
-__pycache__
-*.pyc
